:sparkles:add entity relation tables

+++ src/main/kotlin/com/dev/pierre/projetofinance26062020/controller/ClienteController.kt
@@ -3,7 +3,6 @@ package com.dev.pierre.projetofinance26062020.controller
 import com.dev.pierre.projetofinance26062020.model.Cliente
 import com.dev.pierre.projetofinance26062020.sservice.ClienteService
 import org.springframework.data.domain.Pageable
-import org.springframework.data.jpa.domain.AbstractPersistable_.id
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
@@ -13,19 +12,18 @@ import org.springframework.web.bind.annotation.*
 class ClienteController(val clienteService: ClienteService) {
     
     @PostMapping()
-    fun salaver(@RequestBody cliente: Cliente): ResponseEntity<Unit> {
+    fun salvar(@RequestBody cliente: Cliente): ResponseEntity<Unit> {
         clienteService.craete(cliente)
         return ResponseEntity(Unit,HttpStatus.CREATED)
     }
-    
     @GetMapping("{id}")
-    fun getById(@PathVariable id: Long) : ResponseEntity<Cliente?> {
-        var cliente = clienteService.getById(id)
+    fun findById(@PathVariable id: Int) : ResponseEntity<Cliente?> {
+        var cliente = clienteService.findById(id)
         var status = if (cliente == null) HttpStatus.NOT_FOUND else HttpStatus.OK
         return ResponseEntity(cliente, status)
     }
     @GetMapping
-    fun getAllCliente( pageable: Pageable) =  clienteService.findAllCli(pageable)
+    fun findAll(pageable: Pageable) =  clienteService.findAllCli(pageable)
    /*
     @GetMapping
     fun findAll( pageable: Pageable) : ResponseEntity<Pageable> {
@@ -38,12 +36,22 @@ class ClienteController(val clienteService: ClienteService) {
     }
     */
     @PutMapping("{id}")
-    fun update(@PathVariable id: Long, @RequestBody cliente: Cliente){
-        clienteService.update(id, cliente)
+    fun update(@PathVariable Int: Int, @RequestBody cliente: Cliente): ResponseEntity<Unit>{
+       var status = HttpStatus.NOT_FOUND
+       if (clienteService.findById(Int) != null){
+           clienteService.update(Int, cliente)
+           status = HttpStatus.OK
+       }
+       return ResponseEntity(Unit, status)
     }
     
     @DeleteMapping("{id}")
-    fun deleteCli(@PathVariable id: Long){
-        clienteService.delete(id)
+    fun deleteCli(@PathVariable id: Int): ResponseEntity<Unit>{
+        var status = HttpStatus.NOT_FOUND
+        if (clienteService.findById(id) != null) {
+            status = HttpStatus.OK
+            clienteService.delete(id)
+        }
+       return  ResponseEntity(Unit, status)
     }
 }
